requirements.txt
flask>=3.0.0
authlib>=1.3.0
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
